+++ backend/src/main/java/diyor/adawev/backend/entity/Calculation.java
@@ -3,7 +3,6 @@ package diyor.adawev.backend.entity;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,10 +23,10 @@ public class Calculation {
     private Project project;  // Optional - can be null for calculations without a project
 
     private String roomName;
-    private BigDecimal roomLength;
-    private BigDecimal roomWidth;
-    private BigDecimal roomArea;
-    private BigDecimal pipeLength;
+    private Float roomLength;
+    private Float roomWidth;
+    private Float roomArea;
+    private Float pipeLength;
 
     @Builder.Default
     @OneToMany(mappedBy = "calculation", cascade = CascadeType.ALL)
